app/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

azure-openai==1.3.0
openai==1.3.0
azure-identity==1.14.0

# Vector Database and Embeddings
qdrant-client==2.7.0
sentence-transformers==2.2.2

# HTTP Client
aiohttp==3.9.1
httpx==0.25.2

# Playwright for browser automation
playwright==1.40.0

# Selenium (if needed)
selenium==4.15.2

# JWT and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data processing
pandas==2.1.3

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3

# Development
python-dotenv==1.0.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Retry logic
tenacity==8.2.3

# Logging and monitoring
python-json-logger==2.0.7
